+++ demo/components/victory-line-demo.js
@@ -225,18 +225,19 @@ export default class App extends React.Component {
           label="Hello"
           x={0}
           y={1}
-          theme={VictoryTheme.grayscale}
+          theme={VictoryTheme.material}
         />
 
         <VictoryChart
-          theme={VictoryTheme.grayscale}
+          style={{parent: parentStyle}}
+          theme={VictoryTheme.material}
         >
           <VictoryLine
             style={{parent: parentStyle}}
-            data={this.state.arrayData}
+            data={this.state.data}
             label="Hello"
-            x={0}
-            y={1}
+            x={"x"}
+            y={(d) => (d.y + 15)}
           />
       </VictoryChart>
       </div>