+++ src/components/victory-candlestick/helper-methods.js
@@ -111,18 +111,24 @@ export default {
     }
   },
 
+  isTransparent(attr) {
+    return attr === "none" || attr === "transparent";
+  },
+
   getDataStyles(datum, style, props) {
     const stylesFromData = omit(datum, [
       "x", "y", "size", "name", "label", "open", "close", "high", "low"
     ]);
     const fillCheck = datum.fill || style.fill;
     const strokeCheck = datum.stroke || style.stroke;
-    const strokeColor = fillCheck && !strokeCheck ? fillCheck
-    : strokeCheck;
     const candleColor = datum.open > datum.close ?
             props.candleColors.negative : props.candleColors.positive;
+    const transparentCheck = this.isTransparent(datum.stroke) ||
+    this.isTransparent(style.stroke);
+    const strokeColor = fillCheck || transparentCheck ? fillCheck || candleColor
+    : strokeCheck;
     const baseDataStyle = defaults({}, stylesFromData,
-      {stroke: strokeColor || candleColor, fill: candleColor},
+      {stroke: strokeColor || candleColor, fill: fillCheck || candleColor},
       style);
     return Helpers.evaluateStyle(baseDataStyle, datum);
   },