+++ src/components/victory-errorbar/victory-errorbar.js
@@ -1,57 +1,46 @@
-import { assign, defaults, isFunction, partialRight } from "lodash";
 import React, { PropTypes } from "react";
 import {
   PropTypes as CustomPropTypes, Helpers, Events, VictoryTransition, VictoryLabel,
   VictoryContainer
 } from "victory-core";
-import Bar from "./bar";
-import BarHelpers from "./helper-methods";
+import { defaults, isFunction, partialRight } from "lodash";
+import ErrorBar from "./errorbar";
 import Data from "../../helpers/data";
-import Domain from "../../helpers/domain";
+import ErrorBarHelpers from "./helper-methods";
 
 const fallbackProps = {
   props: {
-    height: 300,
-    width: 450
+    width: 450,
+    height: 300
   },
   style: {
     data: {
-      fill: "#242424",
       opacity: 1,
-      padding: 10,
-      stroke: "transparent",
-      strokeWidth: 0,
-      width: 8
-    },
-    labels: {
-      fill: "#252525",
-      fontFamily: "'Gill Sans', 'Gill Sans MT', 'Ser­avek', 'Trebuchet MS', sans-serif",
-      fontSize: 14,
-      letterSpacing: "0.04em",
-      padding: 10
+      stroke: "#CCC",
+      strokeWidth: 1
     }
   }
 };
 
 const defaultData = [
-  {x: 1, y: 1},
-  {x: 2, y: 2},
-  {x: 3, y: 3},
-  {x: 4, y: 4}
+  {x: 1, y: 1, errorX: 0.1, errorY: 0.1},
+  {x: 2, y: 2, errorX: 0.2, errorY: 0.2},
+  {x: 3, y: 3, errorX: 0.3, errorY: 0.3},
+  {x: 4, y: 4, errorX: 0.4, errorY: 0.4}
 ];
 
-export default class VictoryBar extends React.Component {
-  static role = "bar";
+export default class VictoryErrorBar extends React.Component {
+  static role = "errorBar";
 
   static defaultTransitions = {
     onExit: {
-      duration: 500,
-      before: () => ({ y: 0, yOffset: 0 })
+      duration: 600,
+      before: () => ({ opacity: 0 })
     },
     onEnter: {
-      duration: 500,
-      before: () => ({ y: 0, yOffset: 0 }),
-      after: (datum) => ({ y: datum.y, yOffset: datum.yOffset })
+      duration: 600,
+      before: () => ({ opacity: 0 }),
+      after: (datum) => ({ opacity: datum.opacity || 1 })
     }
   };
 
@@ -78,27 +67,26 @@ export default class VictoryBar extends React.Component {
       })
     ]),
     /**
-     * The data prop specifies the data to be plotted. Data should be in the form of an array
-     * of data points. Each data point may be any format you wish
-     * (depending on the `x` and `y` accessor props), but by default, an object
-     * with x and y properties is expected.
-     * @examples [{x: 1, y: 2}, {x: 2, y: 3}], [[1, 2], [2, 3]],
-     * [[{x: "a", y: 1}, {x: "b", y: 2}], [{x: "a", y: 2}, {x: "b", y: 3}]]
+     * The data prop specifies the data to be plotted.
+     * Data should be in the form of an array of data points and error bar settings.
+     * but by default, an object with x and y properties is expected.
+     * These properties will be interpreted and applied to the individual errorbars
+     * @examples [{x: 1, y: 2, errorX: 0.1, errorY: 0.2}, {x: 2, y: 3, errorX: 0.2, errorY: 0.3}]
      */
+
     data: PropTypes.array,
     /**
-     * The dataComponent prop takes an entire component which will be used to create bars for
+     * The dataComponent prop takes an entire component which will be used to create points for
      * each datum in the chart. The new element created from the passed dataComponent will be
-     * provided with the following properties calculated by VictoryBar: datum, index, scale,
-     * style, events, horizontal (boolean), x, y, and y0. Any of these props may be overridden
-     * by passing in props to the supplied component, or modified or ignored within the custom
-     * component itself. If a dataComponent is not provided, VictoryBar will use its default
-     * Bar component.
+     * provided with the following properties calculated by VictoryErrorBar: datum, index, scale,
+     * style, events, x, y, errorX, errorY. Any of these props may be overridden by passing in
+     * props to the supplied component, or modified or ignored within the custom component itself.
+     * If a dataComponent is not provided, VictoryErrorBar will use its default ErrorBar component.
      */
     dataComponent: PropTypes.element,
     /**
-     * The domain prop describes the range of values your bar chart will cover. This prop can be
-     * given as a array of the minimum and maximum expected values for your bar chart,
+     * The domain prop describes the range of values your chart will include. This prop can be
+     * given as a array of the minimum and maximum expected values for your chart,
      * or as an object that specifies separate arrays for x and y.
      * If this prop is not provided, a domain will be calculated from data, or other
      * available information.
@@ -114,9 +102,9 @@ export default class VictoryBar extends React.Component {
     /**
      * The event prop take an array of event objects. Event objects are composed of
      * a target, an eventKey, and eventHandlers. Targets may be any valid style namespace
-     * for a given component, so "data" and "labels" are all valid targets for VictoryBar events.
-     * The eventKey may optionally be used to select a single element by index rather than an entire
-     * set. The eventHandlers object should be given as an object whose keys are standard
+     * for a given component, so "data" are all valid targets for VictoryErrorBar
+     * events. The eventKey may optionally be used to select a single element by index rather than
+     * an entire set. The eventHandlers object should be given as an object whose keys are standard
      * event names (i.e. onClick) and whose values are event callbacks. The return value
      * of an event handler is used to modify elemnts. The return value should be given
      * as an object or an array of objects with optional target and eventKey keys,
@@ -136,13 +124,7 @@ export default class VictoryBar extends React.Component {
      *            {
      *              eventKey: "theOtherOne",
      *              mutation: (props) => {
-     *                return {style: merge({}, props.style, {fill: "orange"})};
-     *              }
-     *            }, {
-     *              eventKey: "theOtherOne",
-     *              target: "labels",
-     *              mutation: () => {
-     *                return {text: "hey"};
+     *                return {style: merge({}, props.style, {stroke: "orange"})};
      *              }
      *            }
      *          ];
@@ -153,8 +135,9 @@ export default class VictoryBar extends React.Component {
      *}}
      */
     events: PropTypes.arrayOf(PropTypes.shape({
-      target: PropTypes.oneOf(["data", "labels", "parent"]),
+      target: PropTypes.oneOf(["data", "parent"]),
       eventKey: PropTypes.oneOfType([
+        PropTypes.func,
         CustomPropTypes.allOfType([CustomPropTypes.integer, CustomPropTypes.nonNegative]),
         PropTypes.string
       ]),
@@ -187,34 +170,6 @@ export default class VictoryBar extends React.Component {
      */
     height: CustomPropTypes.nonNegative,
     /**
-     * The horizontal prop determines whether the bars will be laid vertically or
-     * horizontally. The bars will be vertical if this prop is false or unspecified,
-     * or horizontal if the prop is set to true.
-     */
-    horizontal: PropTypes.bool,
-    /**
-     * The labels prop defines labels that will appear above each bar in your bar chart.
-     * This prop should be given as an array of values or as a function of data.
-     * If given as an array, the number of elements in the array should be equal to
-     * the length of the data array. Labels may also be added directly to the data object
-     * like data={[{x: 1, y: 1, label: "first"}]}.
-     * @examples: ["spring", "summer", "fall", "winter"], (datum) => datum.title
-     */
-    labels: PropTypes.oneOfType([
-      PropTypes.func,
-      PropTypes.array
-    ]),
-    /**
-     * The labelComponent prop takes in an entire label component which will be used
-     * to create labels for each bar in the bar chart. The new element created from
-     * the passed labelComponent will be supplied with the following properties:
-     * x, y, y0, index, datum, verticalAnchor, textAnchor, angle, style, text, and events.
-     * Any of these props may be overridden by passing in props to the supplied component,
-     * or modified or ignored within the custom component itself. If labelComponent is omitted,
-     * a new VictoryLabel will be created with props described above.
-     */
-    labelComponent: PropTypes.element,
-    /**
      * The padding props specifies the amount of padding in number of pixels between
      * the edge of the chart and any rendered child components. This prop can be given
      * as a number or as an object with padding specified for top, bottom, left
@@ -245,17 +200,17 @@ export default class VictoryBar extends React.Component {
     /**
      * The standalone prop determines whether the component will render a standalone svg
      * or a <g> tag that will be included in an external svg. Set standalone to false to
-     * compose VictoryBar with other components within an enclosing <svg> tag.
+     * compose VictoryErrorBar with other components within an enclosing <svg> tag.
      */
     standalone: PropTypes.bool,
     /**
-     * The style prop specifies styles for your VictoryBar. Any valid inline style properties
+     * The style prop specifies styles for your VictoryErrorBar. Any valid inline style properties
      * will be applied. Height, width, and padding should be specified via the height,
      * width, and padding props, as they are used to calculate the alignment of
      * components within chart. In addition to normal style properties, angle and verticalAnchor
      * may also be specified via the labels object, and they will be passed as props to
-     * VictoryLabel, or any custom labelComponent.
-     * @examples {data: {fill: "red", width: 8}, labels: {fontSize: 12}}
+     * VictoryErrorBar.
+     * @examples {data: {stroke: "red"}}
      */
     style: PropTypes.shape({
       parent: PropTypes.object,
@@ -303,14 +258,14 @@ export default class VictoryBar extends React.Component {
      * The containerComponent prop takes an entire component which will be used to
      * create a container element for standalone charts.
      * The new element created from the passed containerComponent wil be provided with
-     * these props from VictoryBar: height, width, children
+     * these props from VictoryErrorBar: height, width, children
      * (the chart itself) and style. Props that are not provided by the
      * child chart component include title and desc, both of which
      * are intended to add accessibility to Victory components. The more descriptive these props
      * are, the more accessible your data will be for people using screen readers.
      * Any of these props may be overridden by passing in props to the supplied component,
      * or modified or ignored within the custom component itself. If a dataComponent is
-     * not provided, VictoryBar will use the default VictoryContainer component.
+     * not provided, VictoryErrorBar will use the default VictoryContainer component.
      * @example <VictoryContainer title="Chart of Dog Breeds" desc="This chart shows how
      * popular each dog breed is by percentage in Seattle." />
      */
@@ -318,37 +273,45 @@ export default class VictoryBar extends React.Component {
     /**
     * The theme prop takes a style object with nested data, labels, and parent objects.
     * You can create this object yourself, or you can use a theme provided by Victory.
-    * When using VictoryBar as a solo component, implement the theme directly on
-    * VictoryBar. If you are wrapping VictoryBar in VictoryChart, VictoryStack, or
+    * When using VictoryErrorBar as a solo component, implement the theme directly on
+    * VictoryErrorBar. If you are wrapping VictoryErrorBar in VictoryChart, VictoryStack, or
     * VictoryGroup, please call the theme on the outermost wrapper component instead.
     * @example theme={VictoryTheme.grayscale}
     * http://www.github.com/FormidableLabs/victory-core/tree/master/src/victory-theme/grayscale.js
     */
     theme: PropTypes.object,
     /**
-     * The groupComponent prop takes an entire component which will be used to
-     * create group elements for use within container elements. This prop defaults
-     * to a <g> tag on web, and a react-native-svg <G> tag on mobile
+    * The groupComponent prop takes an entire component which will be used to
+    * create group elements for use within container elements. This prop defaults
+    * to a <g> tag on web, and a react-native-svg <G> tag on mobile
+    */
+    groupComponent: PropTypes.element,
+    /**
+     * The borderWidth prop sets the border width of the error bars. `borderWidth` will set
+     * both x, y error bar width.
+     * @type {number}
      */
-    groupComponent: PropTypes.element
+    borderWidth: PropTypes.number
   };
 
   static defaultProps = {
     data: defaultData,
-    dataComponent: <Bar/>,
-    labelComponent: <VictoryLabel/>,
     padding: 50,
     scale: "linear",
     standalone: true,
     x: "x",
     y: "y",
+    borderWidth: 10,
+    dataComponent: <ErrorBar/>,
+    labelComponent: <VictoryLabel/>,
     containerComponent: <VictoryContainer/>,
     groupComponent: <g/>
   };
 
-  static getDomain = Domain.getDomainWithZero.bind(Domain);
+  static getDomain = ErrorBarHelpers.getDomain.bind(ErrorBarHelpers);
   static getData = Data.getData.bind(Data);
-  static getBaseProps = partialRight(BarHelpers.getBaseProps.bind(BarHelpers), fallbackProps);
+  static getBaseProps = partialRight(
+    ErrorBarHelpers.getBaseProps.bind(ErrorBarHelpers), fallbackProps);
 
   constructor() {
     super();
@@ -368,54 +331,39 @@ export default class VictoryBar extends React.Component {
 
   setupEvents(props) {
     const { sharedEvents } = props;
-    this.baseProps = BarHelpers.getBaseProps(props, fallbackProps);
+    this.baseProps = ErrorBarHelpers.getBaseProps(props, fallbackProps);
     this.dataKeys = Object.keys(this.baseProps).filter((key) => key !== "parent");
     this.getSharedEventState = sharedEvents && isFunction(sharedEvents.getEventState) ?
       sharedEvents.getEventState : () => undefined;
   }
 
   renderData(props) {
-    const { dataComponent, labelComponent, groupComponent } = props;
-    const { role } = VictoryBar;
-    const barComponents = [];
-    const barLabelComponents = [];
-    for (let index = 0, len = this.dataKeys.length; index < len; index++) {
-      const key = this.dataKeys[index];
+    const { dataComponent, groupComponent } = props;
+    const errorBarComponents = [];
+    this.dataKeys.forEach((key) => {
       const dataEvents = this.getEvents(props, "data", key);
       const dataProps = defaults(
-        {index, key: `${role}-${key}`, role: `${role}-${index}`},
+        {key: `error-bar-${key}`},
         this.getEventState(key, "data"),
         this.getSharedEventState(key, "data"),
         dataComponent.props,
         this.baseProps[key].data
       );
 
-      barComponents[index] = React.cloneElement(dataComponent, assign(
+      errorBarComponents.push(React.cloneElement(dataComponent, Object.assign(
         {}, dataProps, {events: Events.getPartialEvents(dataEvents, key, dataProps)}
-      ));
+      )));
+    });
 
-      const labelProps = defaults(
-        {key: `${role}-label-${key}`},
-        this.getEventState(key, "labels"),
-        this.getSharedEventState(key, "labels"),
-        labelComponent.props,
-        this.baseProps[key].labels
-      );
+    return React.cloneElement(groupComponent, {}, errorBarComponents);
+  }
 
-      if (labelProps && labelProps.text) {
-        const labelEvents = this.getEvents(props, "labels", key);
-        barLabelComponents[index] = React.cloneElement(labelComponent, assign({
-          events: Events.getPartialEvents(labelEvents, key, labelProps)
-        }, labelProps));
-      }
-    }
-
-    if (barLabelComponents.length > 0) {
-      return React.cloneElement(
-        groupComponent, {}, ...barComponents, ...barLabelComponents
-      );
-    }
-    return barComponents;
+  renderGroup(children, style) {
+    return React.cloneElement(
+      this.props.groupComponent,
+      { role: "presentation", style},
+      children
+    );
   }
 
   renderContainer(props, group) {
@@ -429,18 +377,10 @@ export default class VictoryBar extends React.Component {
     );
     return React.cloneElement(
       props.containerComponent,
-      assign(
+      Object.assign(
         {}, parentProps, {events: Events.getPartialEvents(parentEvents, "parent", parentProps)}
       ),
       group
-    );
-  }
-
-  renderGroup(children, style) {
-    return React.cloneElement(
-      this.props.groupComponent,
-      { role: "presentation", style},
-      children
     );
   }
 
@@ -449,17 +389,21 @@ export default class VictoryBar extends React.Component {
     const { animate, style, standalone } = modifiedProps;
 
     if (animate) {
+      // Do less work by having `VictoryAnimation` tween only values that
+      // make sense to tween. In the future, allow customization of animated
+      // prop whitelist/blacklist?
       const whitelist = [
-        "data", "domain", "height", "padding", "style", "width"
+        "data", "borderWidth"
       ];
       return (
         <VictoryTransition animate={animate} animationWhitelist={whitelist}>
-          {React.createElement(this.constructor, modifiedProps)}
+          {React.createElement(this.constructor, ...modifiedProps)}
         </VictoryTransition>
       );
     }
 
-    const styleObject = modifiedProps.theme && modifiedProps.theme.bar ? modifiedProps.theme.bar
+    const styleObject = modifiedProps.theme && modifiedProps.theme.errorBar
+    ? modifiedProps.theme.errorBar
     : fallbackProps.style;
 
     const baseStyles = Helpers.getStyles(style, styleObject, "auto", "100%");